remove routes refactor
