+++ src/main.rs
@@ -15,10 +15,8 @@ mod auth_handler;
 mod email_service;
 mod errors;
 mod invitation_handler;
-mod invitation_routes;
 mod models;
 mod register_handler;
-mod register_routes;
 mod schema;
 mod utils;
 
@@ -74,12 +72,12 @@ fn main() -> std::io::Result<()> {
                     // routes to invitation
                     .service(
                         web::resource("/invitation")
-                            .route(web::post().to_async(invitation_routes::register_email)),
+                            .route(web::post().to_async(invitation_handler::register_email)),
                     )
                     // routes to register as a user after the
                     .service(
                         web::resource("/register/{invitation_id}")
-                            .route(web::post().to_async(register_routes::register_user)),
+                            .route(web::post().to_async(register_handler::register_user)),
                     ),
             )
             // serve static files
