+++ src/register_handler.rs
@@ -7,6 +7,28 @@ use crate::errors::ServiceError;
 use crate::models::{DbExecutor, Invitation, SlimUser, User};
 use crate::utils::hash_password;
 
+use actix::Addr;
+use actix_web::{web, Error, HttpResponse, ResponseError};
+use futures::Future;
+
+pub fn register_user(
+    invitation_id: web::Path<String>,
+    user_data: web::Json<UserData>,
+    db: web::Data<Addr<DbExecutor>>,
+) -> impl Future<Item = HttpResponse, Error = Error> {
+    let msg = RegisterUser {
+        // into_inner() returns the inner string value from Path
+        invitation_id: invitation_id.into_inner(),
+        password: user_data.password.clone(),
+    };
+
+    db.send(msg)
+        .from_err()
+        .and_then(|db_response| match db_response {
+            Ok(slim_user) => Ok(HttpResponse::Ok().json(slim_user)),
+            Err(service_error) => Ok(service_error.error_response()),
+        })
+}
 // UserData is used to extract data from a post request by the client
 #[derive(Debug, Deserialize)]
 pub struct UserData {
